+++ src/FractalFacade.php
@@ -16,6 +16,6 @@ class FractalFacade extends Facade
      */
     protected static function getFacadeAccessor()
     {
-        return 'laravel-fractal';
+        return 'fractal';
     }
 }